gait-pipeline_openpose/requirements.txt
numpy
pandas
scipy
opencv-python
pyyaml
